Added max-chunk parameter.
The max-chunk parameter is needed for dealing with Singularity container images. Since Singularity does not allow for shared memory to be changed, if the data set is big enough and since max-chunk is 0 by default (load as many as possible as permitted by number of parallel processes), this causes crashes of cubical runs. Specifying max-chunks to a low number (3-4) should allow Cubical to run via Stimela when using Singularity images.